LayoutTests/fast/events/touch/ios/touch-events-when-double-tapping-after-selecting-text-expected.txt

PASS touchEventCount became 4
PASS successfullyParsed is true

TEST COMPLETE
Select me
Then double tap here
Handled touchstart
Handled touchend
Handled touchstart
Handled touchend
